Renamed some variables
